Have to understand gcc -32

+++ 0x00-hello_world/6-size.c
@@ -15,9 +15,10 @@ int integerType;
 long int longintType;
 long long int longlongintType;
 float floatType;
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(charType));
-printf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(integerType));
-printf("Size of a long int: %llu byte(s)\n", (unsigned long) sizeof(longintType));
+
+printf("Size of a char: %lu byte(s)\n", sizeof(charType));
+printf("Size of an int: %lu byte(s)\n", sizeof(integerType));
+printf("Size of a long int: %llu byte(s)\n", sizeof(longintType));
 printf("Size of a long long int: %llu byte(s)\n", sizeof(longlongintType));
 printf("Size of a float: %lu byte(s)\n", sizeof(floatType));
 return (0);
